Add validation metata to proto schema.
This CL introduces a set of annotations that can be used for validating
textpb files. It uses ProtoBuf extensions to carry over extension
metadata describing the type of validation to be performed on a field,
and also any forein key restriction that should be placed on the field.

E.g.: The following change introduced by this CL annotates the
|ad_group_policy| field with the constraints that the field contents
should match the 'name' property of an 'ad_group_policy' type ASSET.

``` proto
    // Describes a set of registry keys that should be applied for a GPO.
    message ActiveDirectoryRegistryPolicy {
      // An identifier for the registry policy.
      string name = 1;

      // The GPO that will be affected by this registry policy.
      string ad_group_policy = 2 [(common.v)={type:ASSET, key:"ad_group_policy"}];
      ...
    }
```

Keeping these validations inline makes it easier to automate validation.
Otherwise any change to the schema in the future would also involve a
costly addition of a set of field validators and reference resolvers.

Bug: none
Change-Id: I07b17ee9ee4d417557399b4c2dd140feedbbeecd
Reviewed-on: https://chromium-review.googlesource.com/710594
Reviewed-by: Asanka Herath <asanka@chromium.org>