Add more robust test framework for mocking GCP REST APIs.
The previous test framework relied on record-replay which would've made
it impossible for someone to introduce a test unless they had access to
the `google.com:chrome-auth-lab-dev` project.

Instead of record-reply based tests which are flawed and likely to
accumulate errors if contributors are excessively careful, this CL
introduces a test framework based on sparsely matching requests and
responding with fixed JSON responses.

The test framework can, but isn't meant to exercise the authentication
layers for generating REST requests to GCP.

Bug:
Change-Id: I0f05a00ed8d53b74a021f1842471625a69b90000
Reviewed-on: https://chromium-review.googlesource.com/795392
Reviewed-by: Asanka Herath <asanka@chromium.org>