Migrate go/gcp to new framework.
In the prototype, the GCP Session object was responsible for loading and
verifying the input configuration. Under the new model, the input files
and assets are independent.

The GCP Session must now be able to function independently of the asset
manifest although it is still expected to know about and use the
HostEnvironment.

Bug:
Change-Id: If9e0056909ed24989ebbb11440bd49b7adeabf45
Reviewed-on: https://chromium-review.googlesource.com/795394
Reviewed-by: Asanka Herath <asanka@chromium.org>