Add a documentation generator for the CEL schema.
The CEL schema is currently documented via comments in the .proto files.
In practice, users shouldn't be expected to pore through .proto files to
figure out out to write a schema. Instead we should generate usable
documentation based on the .proto files.

The version of the documentation generator being checked in should be
considered a framework for something more comprehensive and usable. It
uses the FileDescriptorSets that were emitted during `build.py gen` to
generate MarkDown content describing the package.

Currently this isn't used as a part of `build.py gen`. Things to do
before we can consider this generator as ready:

  * Documentation should include information about the data types and
    formatting for each field. This can automatically be extracted based
    on the `Validation` options.

  * Documentation should include more examples, and have instructions
    for how to add the relevant assets to an asset manifest.

  * There should be a higher level overview document describing how to
    use the generated documentation.

  * We should decide whether to check-in the generated documentation or
    whether to make it available via some other means. Either way, users
    shouldn't need to run the generator or check out the code to view
    the documentation.

Bug:
Change-Id: I81618b3377c291a7a5166dff1d394ad609a79809
Reviewed-on: https://chromium-review.googlesource.com/858122
Reviewed-by: Fei Ling <feiling@chromium.org>