Don't skip over OUTPUT fields in WalkProto()
WalkProto() is a function used to traverse the object tree starting at a
specific proto.Message. It is effectively a depth-first search through
the fields of a proto.Message using Golang/reflect.

Previously, the traversal could be configred to skip over OUTPUT fields.
The latter are a type of field that's expected to be filled in during
deployment (e.g. the email address of a service account that's yet to be
created). By skipping these fields, the traversal could be narrowed to
just those values which were specified in the input.

However, with the advent of RUNTIME fields and inline references, the
WalkProto() abstration can no longer be a determiner of what's input and
what's output. That needs to be resolved at a higher level API that
knows what to do with these new variants of fields. Thus we no longer
need this special case.

This change also includes some method renames.

R=feiling@chromium.org

Change-Id: I2e30db8289177c1203401ef1a8335ada794baa08
Reviewed-on: https://chromium-review.googlesource.com/987154
Reviewed-by: Fei Ling <feiling@chromium.org>