Implement and test Subresource Integrity.
Subresource integrity (https://www.w3.org/TR/SRI/) is used within the
CEL toolchain for implementing content indexed storage as well as
verifying that the stored contents match what was expected. This change
adds a relatively complete SRI implementation and its associated tests.
The tests are based on the a few of the test vectors found in the spec.

R=feiling@chromium.org

Change-Id: I9c4dad5cb1b82f973a4743004a284a4cba703248
Reviewed-on: https://chromium-review.googlesource.com/987133
Reviewed-by: Fei Ling <feiling@chromium.org>