Implement, document, and test a RefPath based trie
The namespace of objects that needs to be processed during
a CEL deployment is organized using a trie. Each objects reference path
is used to index into the trie.

This change includes additional documentation and tests for the trie as
well as support for JSON serialization of the trie contents.

R=feiling@chromium.org

Change-Id: I9c7be6e34d305bc27a93df67b1cc4ec6628be9a6
Reviewed-on: https://chromium-review.googlesource.com/987155
Reviewed-by: Fei Ling <feiling@chromium.org>