Introduce Namespace
The Namespace type represents a hierarchy of proto.Message objects
possibly with internal cross references.

This is the primary mechanism used by CEL for organizing all the resources
that go into a lab. At the heart of the Namespace is a Trie which has a node
corresponding to each and every addressible object and field in the
namespace. Each such node maps to a referenceNode object which contains
additional metadata on the state of that node.

Change-Id: I8594f224db7ebcc3e4d8e8f58ddc55628c8502ba
Reviewed-on: https://chromium-review.googlesource.com/988275
Reviewed-by: Fei Ling <feiling@chromium.org>