Introduce HomomorphicCopy()
This function copies one object to another and works in cases where the
first object isn't strictly assignable to the destination. The function
uses reflection to match similarly named fields and attempt to do the
copy over or convert values from the source to the destination.

It is primarily intended to be a vehicle for transferring data between
Discovery/Schema based generated Go types to ProtoBuf generated Go types
that describes the same schema. Due to both code generators introducing
additional fields to the data structures, the code does not consider it
an error if source and destination doesn't have strict correspondence
between fields.

R=feiling@chromium.org

Change-Id: I41df437793f5bdfcb52465272e96a5e529fadc49
Reviewed-on: https://chromium-review.googlesource.com/988973
Reviewed-by: Fei Ling <feiling@chromium.org>