Fix assignment of OUTPUT values that are proto.Message
Consider something like this:

  type F struct {
    Foo *FooMessage
  }

When publishing a FooMessage at Foo, the Foo field should be updated to
point to the new value being published. This was not happening due to
Namespace calling collectFrom() rather than assign(). This change fixes
the call so that publishing an OUTPUT message works as expected.

It also fixes namespace_node.go to correctly detect when an OUTPUT value
is available upon binding.

R=feiling@chromium.org

Change-Id: I105aba46130466751311bf16fa1f6647963b9f79
Reviewed-on: https://chromium-review.googlesource.com/1025060
Reviewed-by: Fei Ling <feiling@chromium.org>