SRI: Introduce IntegrityLabel
Unfortunately, GCP labels are limited to 63 characters and are further
limited to the token character set. Hence in order to add labels to
objects that represent content digests, those digests need to be encoded
in a format that's consistent with labels.

This CL introduces a IntegrityLabel() method that does that. The
encoding is base-32 and has a prefix that distinguishes it from
compliant SRI tokens.

R=feiling@chromium.org

Change-Id: I186399cf87f3876be9ddb083e3d794f29d0d7064
Reviewed-on: https://chromium-review.googlesource.com/1039925
Reviewed-by: Fei Ling <feiling@chromium.org>