Use URL-safe encoding for embedded subresource integrity tokens.
This change modifies the subresource integrity token to use URL-safe
Base64 encoding. The only difference from the previous encoding is the
use of '-' and '_' in place of '+' and '/'. While not affecting core CEL
functionality, this change means that stored object appear correctly in
GCS browser.

In addition, the ObjectStore interface was previously not documented. It
is now.

R=feiling@chromium.org

Change-Id: I5b7a74b56695cf7716af0c3c2a1abc96d41a1aa7
Reviewed-on: https://chromium-review.googlesource.com/1025061
Reviewed-by: Fei Ling <feiling@chromium.org>