WindowsUser : Raise hardcoded password into WindowsUser.
The previous method for setting a hardcoded password was to embed it in
a the Secret message. This design turns out to be problematic for our
implementation since the password field in WindowsUser is optional.

Since the password field is optional, it can't be marked as OUTPUT.
Since it can't be marked as OUTPUT, it can't be emitted in entirety by
the deployer at deployment time.

Change-Id: Ib06c5842f99666e22590793feb81a659b9349350
Reviewed-on: https://chromium-review.googlesource.com/1042936
Reviewed-by: Fei Ling <feiling@chromium.org>