[cleanup] Another attempt at bringing sanity to long running operations.
Compute engine operations have three further subdivions of operations
based on their scope. For example the client API for querying the status
of operations at the zone, regional, and global scopes are distinct.

In order to prevent an evergrowing list of methods for querying the
status of operations, this CL introduces a single one: JoinOperation().
This applies to all GCP operations.

Within this single function, we dispatch to different (currently three)
handlers based on the schema type of the operation. E.g.:

   JoinOperation(s, o, "Doing foo")

... where `o` is of type *compute.Operation, will result in the
invocation of the compute operation handler. `s` is the *gcp.Session
that's passed around to all CEL GCP APIs.

Furthermore, within the compute operation handler, it will automatically
pick the correct API to invoke based on the operation that's passed in.
Thus the call site can invoke JoinOperation() without knowing the type
nor the scope of the operation.

Change-Id: Iab9252352d5507486cdac82503001d0cc7e32482
Reviewed-on: https://chromium-review.googlesource.com/1045996
Reviewed-by: Fei Ling <feiling@chromium.org>