HomomorphicCopy should deal with mismatched indirection levels.
While generating protobuf messages from discovery document schema, it
turned out that some fields that are optional in the Google API client
libraries can't be marked as opetional in the generated proto files.  It
just so happens that there are no distinguishing attributes for such
fields.

For now, we are going to use an ugly hack here. When copying fields from
"A" to "B", if A has a field that has type *Foo and B has field of type
Foo of the same name, then we'll copy the indirected object from A to B.
The same applies when the indirection exists in reverse.

Change-Id: I6773fec01e9c156b9fadb59cc756e9195cb55be3
Reviewed-on: https://chromium-review.googlesource.com/1045998
Reviewed-by: Fei Ling <feiling@chromium.org>