Add a gcp.ObjectStore method to calculate the absolute GCS path for an object.
This method would be useful for when we need to store a reference that
needs to be evaluated by something that doesn't have the completed asset
manifest. For example, startup scripts.

In these cases, the resolver needs the full GCS path. Object references
normally are relative to a GCS bucket.

Change-Id: Ia396ff6dda38b14d3da9cc0018cbb1abf26c0133
Reviewed-on: https://chromium-review.googlesource.com/1046136
Reviewed-by: Fei Ling <feiling@chromium.org>