Introduce a Getter interface and embed it in common.Context.
Asset resolvers frequently need to poke around the namespace to
determine the properties of neighboring or dependent assets. To make
this possible, this CL introduces a Getter interface and embeds it in
common.Context so that any operation that receives a common.Context will
be able to query it's configuration.

There are additional things we need to do here to make this API
future-proof. For example, during deployment, an asset's resolver should
only be allowed to access other assets that it has declared as its
dependencies. Otherwise the queried values may be race-prone.

Change-Id: Ie462b8d8c3c04374cfdb7668f37ffbfd744525aa
Reviewed-on: https://chromium-review.googlesource.com/1054390
Reviewed-by: Fei Ling <feiling@chromium.org>