Store startup script along with a filename hint.
The GCE Windows Agent requires that the Google Cloud Storage URL to end
with something that resembles a filename in order to figure out what
kind of thing it is. Fortunately, we already built the infrastructure
necessary to store filename hints. We just need to invoke it when
storing the script.

Change-Id: I32b160ae92b928c90f6272bf6ea485af6dedb42c
Bug: 843323
Reviewed-on: https://chromium-review.googlesource.com/1060303
Reviewed-by: Fei Ling <feiling@chromium.org>