Require that newly created disks have autoDelete set to true
When specifying machine types, it is common to specify that a new disk
be created for the new machine. Unless the 'autoDelete' property of this
new disk is set to 'true', these disks aren't cleaned up when the owning
deployment is removed. Thus CEL will fail to completely remove such a
deployment.

In order to work around this, we are now requiring that the 'autoDelete'
property be set to true for newly created disks. It can be left to false
if the machine type is attaching to an existing disk.

Change-Id: Ibf746d8653fe12a815ad4a400f988f9e68041792
Reviewed-on: https://chromium-review.googlesource.com/1060311
Reviewed-by: Fei Ling <feiling@chromium.org>