Better generated documentation for schema.
This change improves the generated schema documentation and adds it to
the list of committed files. Doing so makes the documentation available
to our users without the hassle of generating it themselves.

If changes are made to the .proto files, then the "build.py gen" command
will update the documentation as well.

There are a few more things that should be done WRT documentation:
 * Improve the wording.
 * Add more examples.
 * Exclude Test messages and their fields from the documentation.
 * Test the documentation with some of our customers to make sure they
   can understand it.

Bug: none
Change-Id: Ibbfa551c762b73d3a973b4080b42cae594414c98
Reviewed-on: https://chromium-review.googlesource.com/1061616
Reviewed-by: Fei Ling <feiling@chromium.org>