Raise Exception when _CheckAndInstall fails.
We silently succeed `build.py deps --install` and create the sentinel file when
we fail to run/install components. When `build.py build` inevitably fails, it
suggests that the GOBIN path might not have been set and to rerun
`build.py deps --install`, but we then jump over `deps ensure` & creating the
third_party directory because of the sentinel file.

Raising an Exception also makes it clear that the installation failed.

+ Fixed minor typos

Change-Id: I60cba43141b23239fd5e49845df75f899b1de357
Reviewed-on: https://chromium-review.googlesource.com/c/1255557
Reviewed-by: Fei Ling <feiling@chromium.org>