Extend Remote Desktop support on Win2012.
`create_remote_desktop_win2012.ps1` doesn't work on non-DC computers because
the local admin account which we use for setup is not active.

Use the domain admin instead of the local admin so RDS can be configured on
any machine in a domain. The RDS infrastructure requires servers to be joined
to a domain.

Also moved the RDS configuration to our DSC using the xRemoteDesktopSessionHost
module. v1.8.0.0 has a fix to a bug that affected us before (issue #39).

Verified RDS (visited /rdweb & launched calc.exe) on these asset configurations:
1) RDS on DC (2012), visited from client
2) RDS on client (2012), visited from DC
3) RDS on both DC and client (2012)

== Deploy on a domain controller or not?
The industry best practices posts I've read on deploying Prod RDS suggested
that deploying it on a DC was a bad idea for perf and security reasons.

Microsoft also recommends not to install RDS on a DC:
2008 : https://docs.microsoft.com/en-us/previous-versions/windows/it-pro/
       windows-server-2008-R2-and-2008/cc742817(v=ws.11)
> Installing the RD Session Host role service on an Active Directory domain
> controller is not recommended. Allowing users to run programs on a domain
> controller could create security risks and performance issues.

2016, 2019 : https://support.microsoft.com/en-us/help/2833839/guidelines-for-
             installing-the-remote-desktop-session-host-role-service
> This configuration should be used only when it is the only option. This is
> because the recommended configuration includes the Remote Desktop Connection
> Broker to provide access to the complete functionality with Remote Desktop
> Services. If a Domain Controller (DC) is available on a separate server, we
> recommend that you use the Standard Remote Desktop Services deployment wizard.
> This configuration is appropriate when there is only one server on the network.

Change-Id: I0eafe6e8e1b7ecc407ebfce8137248d4b36721b9
Reviewed-on: https://chromium-review.googlesource.com/c/1277879
Reviewed-by: Fei Ling <feiling@chromium.org>