Fetch deterministic protobuf packages (fixes bots)
Context:
Our bots started failing a couple of days ago after proto-gen-go
changes were checked in to github.

This broke many people including us.
https://github.com/golang/protobuf/issues/763

This highlighted a known problem we have with deterministic builds -
we fetch a lot of unversioned packages (straight from their master)
which means that we are not only vulnerable to breaking changes, but
also that those changes can break previously good CELab commits.

We should eventually do that for all of our packages, but I've only
taken care of protobuf here.

Change-Id: I0facc895234f3247bee1202e9647e2dc4355518d
Reviewed-on: https://chromium-review.googlesource.com/c/1356765
Reviewed-by: Fei Ling <feiling@chromium.org>