Improve ActiveDirectory E2E tests
Added some assertions in our ActiveDirectory tests and removed the last test
which wasn't testing anything the other two aren't already.

Also found out why the 3rd test's asset file was failing:
We don't support having the same user in different domains (`joe`).
This is clearly a bug on our side but it doesn't seem too bad for now.
We can fix it later if it becomes a problem.

Change-Id: Ia1bfdcf75eae2b19edee6953c7051929b2b2f998
Reviewed-on: https://chromium-review.googlesource.com/c/1387105
Reviewed-by: Fei Ling <feiling@chromium.org>