Implement SharedHostProvider in ./run_tests.py
This allows us to share projects between different test runs.
SharedHostProvider will only return hosts for which it has a lock for.
We release the lock when we no longer need the projects.
Locks are also ignored when they are over an hour old (we refresh them every 20 mins).
This is to prevent catastrophic failures in run_tests to leave projects locked forever.

Change-Id: Icfb850edd01f00582d479bc395a0af855a71eb1b
Reviewed-on: https://chromium-review.googlesource.com/c/1392510
Reviewed-by: Fei Ling <feiling@chromium.org>