Improve _BuildStep logging in build.py
We've seen intermittent build failures on our try-bots.
(none on our CI bots, but that might just be luck)

The issue appears and disappears depending on the day and fails the build with:
> go\gcp\deploy\builtins.go:19:38: undefined: _escFSMustByte

This function is in resources.gen.go that we should generate before this step.

There's some stamping/timestamp logic that can skip the generation step and I
suspect there's a bug there when reusing old bots (can't repro locally).
However, the logging is insufficient to know why from LUCI logs and I couldn't
get a repro today (of course...).

Changes:
  Added more logging to help investigate this the next time it happens.

  No change for normal `build.py build`, changes only `--verbose` builds.
  Our recipe uses `--verbose`.

  Also fixed formatting of long lines to silence warnings.

Change-Id: Ifc86b3322c2fbb532def1c391c7b5bad6526fdec
Reviewed-on: https://chromium-review.googlesource.com/c/1443640
Reviewed-by: Fei Ling <feiling@chromium.org>