Save compute logs when cel_ctl fails in test.py
We used to wait for onhost configuration in test.py after cel_ctl returned but
recently moved this code to `cel_ctl deploy`:
https://chromium-review.googlesource.com/c/enterprise/cel/+/1388185

Because there was no compute logs to really save when `cel_ctl deploy` failed,
the original test.py code skipped saving logs when it did (cases 1 & 2 below).

With the recent change to move the WaitForAllAssetsReady code to `cel_ctl`,
there can now be useful logs to save when it happens.

Change:
  Always upload logs when `cel_ctl deploy` fails.
  There are 3 main phases in `cel_ctl deploy`:
  1- Clean up project: If it fails here, we could get logs from old instances.
       It'd be pretty surprising to see this in LUCI CI tests because we clean up
       the environment at the end of each test. The previous test would've had to
       also fail delete, which is definately possible, but this situation is easy
       enough to spot from the MILO test logs.
  2- Create resources/deployment: If it fails here, we'd have no instance and
       WriteComputeLogs would no-op. Perfect.
  3- WaitForAllAssetsReady: We definitely want logs here.

Example of logs not being uploaded during WaitForAllAssetsReady failure:
https://ci.chromium.org/p/celab/builders/luci.celab.ci/Linux/b8922792479035364496

Change-Id: I587509cd2468cd18269dcf164b7f9ae0dc9e4bea
Reviewed-on: https://chromium-review.googlesource.com/c/1448851
Reviewed-by: Fei Ling <feiling@chromium.org>