Create new Daily LUCI bot
This is a follow-up to "Automatically run our test try bot every morning.":
https://chromium-review.googlesource.com/c/enterprise/cel/+/1506279

Turns out that the scheduler doesn't have permissions to queue builds in our try bucket,
so let's use our CI bucket instead.

Also added the new bot to our list of notifications so we get an email when it fails.

About:
>We kick off try bots instead of CI bots to avoid uploading binaries to Google Storage.

With this new bot name, they'd end up in a separate directory, so I'm not too concerned about this anymore.

Change-Id: Idaea35483f34fdb468392150a696cea022f8d13f
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1509172
Reviewed-by: Fei Ling <feiling@chromium.org>