Fix intermittent RDS setup broker connection failures (Win2012).
We're seeing intermittent failures on RDS setup (Win2012):
```
Validation failed for the "RD Connection Broker" parameter.
win2012-client.domain2012.com    Unable to connect to the server by using Windows PowerShell remoting. Verify that you
can connect to the server.
    + CategoryInfo          : NotSpecified: (:) [], CimException
    + FullyQualifiedErrorId : Microsoft.PowerShell.Commands.WriteErrorException,New-RDSessionDeployment
    + PSComputerName        : localhost
```

Got a local repro and was able to log in, and rerun the dsc config a bunch of times.
I kept hitting that error on [xRDSessionDeployment]Deployment.

Adding RDS-Connection-Broker to the DependsOn list did the trick.

Change-Id: I2238c007ae9dd986518292addb5b9420f6b19b29
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1525131
Reviewed-by: Fei Ling <feiling@chromium.org>