Extend long running operation timeout with JoinOperationWithTimeout.
When DeleteObsoleteDeployments or InvokeDeploymentManager times out,
it's devastating for our tests because we will try to reuse that project
for another test and will fail to deploy (because another operation is running).

This can lead to the rest of the pending tests failing after trying to use
the "corrupted" project until the operation is over.

Example:
https://logs.chromium.org/logs/celab/buildbucket/cr-buildbucket.appspot.com/8918340443520421136/+/steps/run_all_tests/0/stdout

Change-Id: I379994ff0dbf10b3dd35f8ac6c5b24e79ab30aea
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1534968
Reviewed-by: Fei Ling <feiling@chromium.org>