Don't fail asset when script exits because of a machine restart.
This is one of our top intermittent failures right now.

We try to control when we reboot with DSC `RebootNodeIfNeeded = $false`,
but this doesn't apply for configurations that require more than one reboot,
for example, setting up RemoteDesktop. In those scenarios, our LCM picks the
config back up on boot and can restart while we're processing another script.

There's a new property for DSC called "ActionAfterReboot", but it doesn't
seem to always be honored (but might be because of other things we do/install).

For this reason, I took a different approach and added code to forgive
failures that happen when Windows is shutting down.

Verified:
- Legitimate failure still marks the asset as failed
- Failure during shutdown/restart doesn't mark the asset as failed, the machine restarts and continues the configuration

Change-Id: I2a662898c95be00264ec30d66251e6b8607fcbdc
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1521666
Reviewed-by: Fei Ling <feiling@chromium.org>