Disable Windows Update on NestedVMs.
Our NestedVMs sometime fail to restart (~10%) after Rename-Computer, which
means our tests fail because the machine name isn't updated.

Looking into it further, it seems Windows is stuck on the "Restarting..." step.

Event logs are not very helpful either - just silence after receiving the
shutdown signal, no errors.

After researching this a bit, it seems this could be caused by Windows Update
hanging, so I tried to stop the service and that failed about 10% of the time.
That 10% were also instances in which we hang on "Restarting...".

Added code to force kill Windows Update when it doesn't want to stop.

Seems to do the trick - each of those runs is a 10x NestedVM test:
https://ci.chromium.org/p/celab/builders/try/tests/b8916854394150487984 // unrelated failure
https://ci.chromium.org/p/celab/builders/try/tests/b8916848606918694496
https://ci.chromium.org/p/celab/builders/try/tests/b8916805527896518080
https://ci.chromium.org/p/celab/builders/try/tests/b8916801503262491568
https://ci.chromium.org/p/celab/builders/try/tests/b8916791198566155952

Also:
- Merged the two Deploy methods so they have a similar steps order because
  it was confusing to keep track of when different steps (e.g. Prepare)
  were being called on nested VMs vs normal instances.
- Slightly improved logging.
Change-Id: Id350a86bbe3767c2a74eee6d8aa730a000a720a0
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1557582
Reviewed-by: Fei Ling <feiling@chromium.org>