Add flag --chromedriver to support using the chromedriver built from the source.
Another change is that in SetPolicy(), we set the policy for both Chrome & Chromium.
Since the chrome built from the source most likey is Chromium, not Chrome.
Only googlers can build Chrome branded executable, and that also requires some setup.
So better to just support both.

I've thought about of detecting whether the executable is Chrome or Chromium. But there is
no easy way to do that. Plus, I don't see any problems of just supporting both.

The safe_search_webdriver_test now has code to run on logging for both chrome & chromedriver.
This can serve as an example for people who needs to investige problems.

Change-Id: Ideeb7c263550f3e0b88c1aa0fe3e2590e536b8b4
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1622049
Reviewed-by: Xiang Ji <jxiang@google.com>
Reviewed-by: Mathieu Binette <mbinette@google.com>