Add enterprise test for TranslateEnabled policy.
The key thing that is different from the current Manual steps is the verification part:
> 3A. Verify Translate pop up appears.
> 4A. Verify Translate pop up should not be displayed.

WebDriver doesn't offer features to read/interact with browser notifications,
so the workaround I use is to automatically accept translate notifications
and look at the text on the page to detect if a translation occurred or not.

If you have better suggestions, I'm listening! :)

Change-Id: Ibf7054c9d9a4c973782a0da01561eafb8f32226f
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1628816
Reviewed-by: Fei Ling <feiling@chromium.org>