Support different test configs on linux & windows.
This adds support for launching tests on Windows builders (after fixes in https://crrev.com/c/1642366) and creates different try builders for different test configs.

Added a "quick tests" mixin that launches 2 tests (the minimum required to exercise all the lock & threading logic) and is useful for testing generic test infra changes. It launches two of our faster tests which takes ~20mins instead of an hour. We still run all tests (full-tests) in our Linux CI and Daily run.

This is associated with this recipe change:
https://crrev.com/c/1642648

Change-Id: I036f31627f4f26607d66975345e9f9d5b4aed38f
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1641755
Reviewed-by: Fei Ling <feiling@chromium.org>