Replace RunCommand's StackDriver logging with GetSerialPortOutput
We originally used StackDriver for RunCommand output because of the nice API:
- Ability to label log entries with custom information;
- Ability to query log entries based on those tags/operation information.

The API looks great in theory, but in practice, there were annoying limitations:
1- High latency:
   It can take 5-10 seconds between a log.write() and cel_ctl being able to
   query it from the API.
2- Read quotas:
   The logging API has project-wide read quotas of 1/s, which limits our
   ability to execute commands on multiple instances in parallel.
3- Non-deterministic visibility order of entries:
   See https://crrev.com/c/1457306

The main disadvantage of SerialPortOutput is the lack of label/query features.
I replaced it with a line format that is less robust than the former approach.

That said, some of the advantages of using GetSerialPortOutput:
1) Lower latency:
ex:
```
RunPowershell("""
  for ($i = 0; $i -lt %s; $i++) {
    Write-Host "This is line #$i"
  }
""")
```

This script runs in an average of 4.5 seconds VS 12 seconds before.
The command execution time is 2-3s according to console timestamps.
That's a significant cut of the `run` cmd overhead (10s to 2s).

2) No quotas

3) Deterministic line order:
What we have in instance logs is exactly what was read, there's only one feed.
Easy to debug/investigate.

Change-Id: I9ed3893ae9ccba7f02e847aa7d91c7f43dce6340
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1457486
Reviewed-by: Fei Ling <feiling@chromium.org>