Install PackageManagement_x64.msi on all Windows machines.
New Windows Server 2012R2 GCE images don't have PackageManagement installed.
https://cloud.google.com/compute/docs/images#windows_server

I'm not sure if that's by-design or a bug (it used to be installed before),
but it doesn't matter - let's just install it ourselves on all machines.

We already did this on Windows Server 2008, so I copied the code to
prepare_win2012.ps1, and realized the 2008 and 2012 scripts were nearly
identical, so I refactored the common parts to prepare_windows.ps1 and
got rid of prepare_win2012.ps1.

Validation:
https://cr-buildbucket.appspot.com/build/8902114951489356864
https://cr-buildbucket.appspot.com/build/8902114951489356848
[this doesn't fix UITest -- https://crrev.com/c/1804276 does]

Change-Id: Ib182c1da3e1ffa501d29c71a5ed258217cae7fd9
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1808402
Reviewed-by: Fei Ling <feiling@chromium.org>