Change quick-tests builder mixin to categories.
The `quick-tests` builders are quick versions of celab bots that run
a subset of tests that cover generic test infra changes more quickly
than a full celab test run and were introduced a couple of months ago
in https://crrev.com/c/1641755.

Switching the list of tests there to categories to test out the new
system that we'll be using in chromium/src for chrome-only tests.

Also took the opportunity to add UITest to the list of "core" tests.

Requires recipe change in https://crrev.com/c/1835668.

Tested with:
https://ci.chromium.org/raw/build/logs.chromium.org/celab/led/mbinette_google.com/c964e05593c6d97e61381be3c0be9c7a50a1109b12951418bb04d9c897bb4181/+/annotations?server=chromium-swarm.appspot.com

Change-Id: I4631ef7ab528382066701e9864519ff25c9e38bb
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1835025
Reviewed-by: Fei Ling <feiling@chromium.org>