Workaround the problem that Win10 VM cannot boot sometimes.
When this happens, my investigation showed that almost certainly the VM freezes and is
totally unresponsive. Killing & restarting KVM works most the time (I haven't seen any
cases when this does not work). I know this will not work 100%, but it will greatly reduce
the failure rate.

Also mark the user_name & password as required for nested VM.

Testing: I tested the fix 50 times. The fix (kill then restart qemu/kvm)
was called 4 times, and the fix worked every time it was called.


Change-Id: I97d47cfb2fad54188ec8a65971c31927452625ab
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1925680
Reviewed-by: Mathieu Binette <mbinette@google.com>