[CELAB] Updating the docs & one build step
Updating the docs to be a little more explicit on installing the protos.
Additionally newer proto versions are not "v0.beta", so I changed that
to dynamically find the folder name.

This is not the best solution, but highlights the issue and a possible
solution. I am open to suggestions on a more proper solution, and I
believe we should change this across the build file to all applicable
locations

TEST=python build.py build
BUG=None

Change-Id: Ic10bb392854def00530f2deb5cd891dc813cfac4
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1967789
Reviewed-by: Mathieu Binette <mbinette@google.com>