Move default_client.go (and luci/auth token source) to cel_ctl package.
We only use it in cel_ctl, but it's currently defined in go/gcp, which is
a package also used to build cel_agent.

That isn't a problem here, because luci/auth supports Windows externally,
but it doesn't internally (by-design), and we "need" it to build Windows
cel_agent.exe.

You can ping me for more details.

To get around this, I moved `luci/auth` references to the cel_ctl package,
so we no longer need `luci/auth` to build cel_agent.

I also considered fixing internal Windows builds for the luci/auth package,
but LUCI would have to agree to support it internally moving forward and
it's not worth it since we don't even need it outside of cel_ctl
(which we only need on Linux internally -- and not even on LUCI!).

Change-Id: I719fce5f670f646e54af4c4d7efd5fe292e75a80
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1976682
Reviewed-by: Xiang Ji <jxiang@google.com>