Minor changes to support Python3 builds & tests.
We're eventually going to have to move to Python3 when our deps drop
PY2 support, so thought I'd take a quick look at what's failing on PY3.

I will also send out a doc with the full list of things that are still
missing (a few py3 vpython wheels, etc).

This is not an exhaustive CL for full PY3 support, just some
low-hanging fruits.

Changes work on both PY2 and PY3.

Validation:
> python3 build.py build

Change-Id: I88415232f1e1fbaa7db30326508a223f68622ed3
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1988707
Reviewed-by: Derek Beckett <dbeckett@chromium.org>