Update test pool configuration scripts.
Using `gcloud functions deploy` now prompts for an answer to
"allow anonymous callers". Using `--quiet` will skip the prompt
and use the default value (N), which is what we want and used to be
the old default.

Also added `servicemanagement` to the list of services to warm-up --
cel_ctl would try do it the first time as well, but some test service
accounts might not have the permission to enable APIs, and we try to
start them up here anyway to save time.

Change-Id: I18a3ef7b082c2bafb764814694c4b27854311f18
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/1990313
Reviewed-by: Fei Ling <feiling@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>