Install celab test client deps offline.
As part of setting up test client machines, celab downloads third-party dependencies from the
Internet. The download and installation of 3rd-party deps is a frequent
failure point, as the dep origin link experiences network downtime and
as the online deps change. This change caches 3rd-party deps on CIPD, and downloads
cached versions of deps during installation, thereby avoiding the
network failure point and giving celab control on what binaries it
installs on the test client.

Change-Id: I9a91bde4bf368e1cd221478b154370fc6b4a9a70
Reviewed-on: https://chromium-review.googlesource.com/c/enterprise/cel/+/2048217
Reviewed-by: Xiang Ji <jxiang@google.com>